#5 android-specific Screen API moved to clean Fragment API version
