+++ widgets/src/androidMain/kotlin/dev/icerock/moko/widgets/screen/DrawerNavigationScreen.kt
@@ -4,17 +4,9 @@
 
 package dev.icerock.moko.widgets.screen
 
-import android.content.Context
-import android.view.ViewGroup
-import dev.icerock.moko.widgets.core.View
-
 actual abstract class DrawerNavigationScreen actual constructor(
     screenFactory: ScreenFactory
 ) : Screen<Args.Empty>() {
-    override fun createView(context: Context, parent: ViewGroup?): View {
-        TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
-    }
-
     actual abstract val header: Any?
     actual abstract val items: List<DrawerNavigationItem>
