+++ widgets/src/androidMain/kotlin/dev/icerock/moko/widgets/screen/Screen.kt
@@ -4,18 +4,13 @@
 
 package dev.icerock.moko.widgets.screen
 
-import android.content.Context
-import android.os.Bundle
 import android.os.Handler
 import android.os.Looper
-import android.view.LayoutInflater
-import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
 import dev.icerock.moko.mvvm.createViewModelFactory
 import dev.icerock.moko.mvvm.dispatcher.EventsDispatcher
 import dev.icerock.moko.mvvm.viewmodel.ViewModel
-import dev.icerock.moko.widgets.core.View
 import java.util.concurrent.Executor
 
 actual abstract class Screen<Arg : Args> : Fragment() {
@@ -38,14 +33,4 @@ actual abstract class Screen<Arg : Args> : Fragment() {
         get() {
             return parentFragment as? Screen<*>
         }
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): android.view.View? {
-        return createView(context = requireContext(), parent = container)
-    }
-
-    abstract fun createView(context: Context, parent: ViewGroup?): View
 }
