+++ widgets/src/androidMain/kotlin/dev/icerock/moko/widgets/screen/WidgetScreen.kt
@@ -5,11 +5,11 @@
 package dev.icerock.moko.widgets.screen
 
 import android.app.Service
-import android.content.Context
+import android.os.Bundle
+import android.view.LayoutInflater
 import android.view.ViewGroup
 import android.view.WindowManager
 import android.view.inputmethod.InputMethodManager
-import dev.icerock.moko.widgets.core.View
 import dev.icerock.moko.widgets.core.ViewFactoryContext
 import dev.icerock.moko.widgets.core.Widget
 import dev.icerock.moko.widgets.style.view.SizeSpec
@@ -18,13 +18,17 @@ import dev.icerock.moko.widgets.style.view.WidgetSize
 actual abstract class WidgetScreen<Arg : Args> actual constructor() : Screen<Arg>() {
     actual abstract fun createContentWidget(): Widget<WidgetSize.Const<SizeSpec.AsParent, SizeSpec.AsParent>>
 
-    override fun createView(context: Context, parent: ViewGroup?): View {
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): android.view.View? {
         val widget = createContentWidget()
         val view = widget.buildView(
             ViewFactoryContext(
-                context = context,
+                context = requireContext(),
                 lifecycleOwner = this,
-                parent = parent
+                parent = container
             )
         ).view
         // TODO support margins?
